+++ build.rs
@@ -1,10 +1,7 @@
 use rustc_version::{version_meta, Channel};
 
 fn main() {
-    match version_meta().unwrap().channel {
-        Channel::Nightly => {
-            println!("cargo:rustc-cfg=nightly_build");
-        }
-        _ => {}
+    if version_meta().unwrap().channel == Channel::Nightly {
+        println!("cargo:rustc-cfg=nightly_build");
     }
 }
